move fibonacci-air crate and use it as aa test for public values
