--- fibnonacci-air/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "fibnonacci-air"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-p3-air = { path = "../air" }
-p3-field = { path = "../field" }
-p3-matrix = { path = "../matrix" }
-p3-util = { path = "../util" }
-serde_json = "1.0.114"
-
-[dev-dependencies]
-p3-baby-bear = { path = "../baby-bear" }
-p3-challenger = { path = "../challenger" }
-p3-commit = { path = "../commit" }
-p3-dft = { path = "../dft" }
-p3-fri = { path = "../fri" }
-p3-merkle-tree = { path = "../merkle-tree" }
-p3-poseidon = {path = "../poseidon"}
-p3-poseidon2 = { path = "../poseidon2" }
-p3-symmetric = { path = "../symmetric" }
-p3-uni-stark = { path = "../uni-stark" }
-rand = "0.8.5"
-
-[[example]]
-name = "prove_baby_bear_poseidon2"
